api/requirements.txt
fastapi
uvicorn
motor
pymongo[snappy,gssapi,srv,tls]
python-multipart
python-dotenv
python-jose
httpx
pydantic
